refactor(v2): rename modules, establish pyrcel.legacy subpackage (#33)
Move all NumPy/numba legacy implementations under pyrcel/legacy/ and give the
JAX v2 modules clean top-level names (thermo, parcel_aux, integrator, equilibrate,
model). Create pyrcel/activation.py shim re-exporting from legacy pending the JAX
rewrite in #36. Update all internal imports, tests, pyproject.toml per-file-ignores,
and pyrcel/__init__.py public API surface. 114 fast tests pass.
