+++ pyproject.toml
@@ -80,21 +80,28 @@ exclude = ["docs/examples/*.ipynb"]
 select = ["E", "F", "I", "UP", "NPY"]
 
 [tool.ruff.lint.per-file-ignores]
-# --- legacy files (moving to pyrcel.legacy in #33) ---
-# Violations here will be resolved after the module rename. Any suppression
-# that survives #33 must be explicitly audited and documented — tracked in #51.
-"pyrcel/thermo.py" = ["F403", "F405", "UP031", "E731", "E501"]
-"pyrcel/parcel.py" = ["UP031", "E731", "F841", "E501"]
-"pyrcel/activation.py" = ["UP031", "E731", "F841", "E501"]
-"pyrcel/_parcel_aux_numba.py" = ["UP031", "F841", "E501"]
+# --- pyrcel.legacy files ---
+# These are the preserved NumPy/numba reference implementations. Violations here are
+# tracked for future cleanup in #51. Any suppression that survives that audit must be
+# explicitly documented.
+"pyrcel/legacy/thermo.py" = ["F403", "F405", "UP031", "E731", "E501"]
+"pyrcel/legacy/parcel.py" = ["UP031", "E731", "F841", "E501"]
+"pyrcel/legacy/activation.py" = ["UP031", "E731", "F841", "E501"]
+"pyrcel/legacy/parcel_aux.py" = ["UP031", "F841", "E501"]
+"pyrcel/legacy/integrator.py" = ["E731", "F841", "UP031", "UP036", "F401", "E501"]
+"pyrcel/legacy/driver.py" = ["UP031", "F841", "E501"]
+"pyrcel/legacy/postprocess.py" = ["UP031", "E731", "F841", "E501"]
+"pyrcel/legacy/vis.py" = ["E731", "F841", "UP031", "E501"]
+# --- public API shims ---
+# F401: explicit `X as X` re-exports are intentional public API; I001: ruff isort
+# rejects grouped re-export blocks — suppress rather than let it split them per-symbol.
+"pyrcel/__init__.py" = ["F401", "I001"]
+"pyrcel/activation.py" = ["F401"]
+# --- other legacy/non-JAX package files ---
 "pyrcel/aerosol.py" = ["UP031", "E731", "F841", "E501"]
 "pyrcel/distributions.py" = ["UP031", "E501"]
-"pyrcel/integrator.py" = ["E731", "F841", "UP031", "UP036", "F401", "E501"]
-"pyrcel/driver.py" = ["UP031", "F841", "E501"]
 "pyrcel/output.py" = ["UP031", "E731", "F841", "E501"]
-"pyrcel/vis.py" = ["E731", "F841", "UP031", "E501"]
 "pyrcel/scripts/*.py" = ["UP031", "F841", "E501"]
-"pyrcel/__init__.py" = ["F403", "F405", "F401", "E501"]
 "pyrcel/test/*.py" = ["F403", "F405", "F401", "UP031", "E501"]
 "pyrcel/constants.py" = ["E501"]
 # --- non-package files ---
@@ -107,5 +114,5 @@ markers = [
     "slow: full JAX/diffrax integration, autodiff, or ensemble solve (skip locally with -m 'not slow')",
 ]
 filterwarnings = [
-    "ignore:invalid value encountered in scalar divide:RuntimeWarning:pyrcel.activation",
+    "ignore:invalid value encountered in scalar divide:RuntimeWarning:pyrcel.legacy.activation",
 ]
